# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de fotografias e eventos caninos do PS (Photo Storage).",
        "title": "PS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte PS",
            "email": "suporte@ps.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/people": {
            "get": {
                "description": "Retorna a lista de pessoas do tenant autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Listar pessoas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/person.Person"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Cadastra uma nova pessoa (cadastro único) para o tenant autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Criar pessoa",
                "parameters": [
                    {
                        "description": "Dados da Pessoa",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/people/{id}": {
            "get": {
                "description": "Retorna os detalhes de uma pessoa específica pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Obter pessoa por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Pessoa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de uma pessoa existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Atualizar pessoa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Pessoa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da Pessoa",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/person.Person"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove uma pessoa do tenant autenticado",
                "tags": [
                    "People"
                ],
                "summary": "Excluir pessoa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Pessoa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "person.Person": {
            "type": "object",
            "properties": {
                "alternative_email": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}